jobs/requirements.txt
boto3
panda
bs4